requirements.txt
# Core Django
asgiref==3.9.1
Django==5.2.5
sqlparse==0.5.3

# Production server
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Environment management
python-decouple==3.8

# API framework
djangorestframework==3.15.2

# CORS handling
django-cors-headers==4.3.1

# AI and language processing
openai==1.51.0
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7

# Twilio integration
twilio==9.0.0

# HTTP requests
requests==2.31.0

# Image processing
Pillow==10.4.0

# Caching (for production)
redis==5.0.1
django-redis==5.4.0

# Background tasks (for production)
celery==5.3.6

# WebSockets (optional)
channels==4.1.0
channels-redis==4.2.0

# Monitoring (optional)
sentry-sdk==2.15.0

# Media storage (for production)
boto3==1.35.0
django-storages==1.14.2

# Security
django-ratelimit==4.1.0
